pyproject.toml
[tool.poetry]
name = "xrvoyage"
version = "0.0.0"
description = "XR Voyage Python Package"
authors = ["Peter Styk <peter@neux.io>"]
packages = [
    { include = "xrvoyage" },
]
[tool.poetry.dependencies]
python = ">=3.11.0"
pydantic = "^2.7.1"
python-decouple = "^3.8"
requests = "^2.32.2"
logzero = "^1.7.0"
websockets = "^12.0"
pyjwt = "^2.8.0"

lionagi = "^0.2.1"
[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=42", "wheel", "setuptools_scm"]
build-backend = "poetry.core.masonry.api"

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"
write_to = "xrvoyage/_version.py"

[tool.poetry.dev-dependencies]
pytest = ">=6.2,<7.0"
pytest-asyncio = "^0.15.1"
python-dotenv = "^1.0.0"
pytest-dependency = "^0.6.0"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests"
]